# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAdmin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IQueryParam": {
				"properties": {
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"isSpecial": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IEventInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isSpecial": {
						"type": "boolean"
					},
					"thumbnail": {
						"type": "string"
					},
					"admin_id": {
						"type": "string"
					},
					"isdeleted": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"time",
					"description",
					"thumbnail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ievent": {
				"properties": {
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"isdeleted": {
						"type": "boolean"
					},
					"event_info": {
						"items": {
							"$ref": "#/components/schemas/IEventInfo"
						},
						"type": "array"
					}
				},
				"required": [
					"date",
					"event_info"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITable": {
				"properties": {
					"table_no": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"isdeleted": {
						"type": "boolean"
					}
				},
				"required": [
					"table_no",
					"capacity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phonenumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profile": {
						"type": "string"
					},
					"isdeleted": {
						"type": "boolean"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"phonenumber",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IuserLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "easy-chill",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/auth/admin": {
			"post": {
				"operationId": "AdminSignup",
				"responses": {
					"201": {
						"description": "Create Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAdmin"
							}
						}
					}
				}
			}
		},
		"/auth/admin-login": {
			"post": {
				"operationId": "AdminLogin",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAdmin"
							}
						}
					}
				}
			}
		},
		"/auth/admin/{id}": {
			"get": {
				"operationId": "GetAdminInfo",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/event": {
			"get": {
				"operationId": "GetAllEvent",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isSpecial",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"201": {
						"description": "Created Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Ievent"
							}
						}
					}
				}
			}
		},
		"/event/{id}": {
			"get": {
				"operationId": "GetEventById",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEvent",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Ievent"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEvent",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/table": {
			"post": {
				"operationId": "CreateTable",
				"responses": {
					"201": {
						"description": "Created Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITable"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllTable",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": []
			}
		},
		"/table/{tableId}": {
			"get": {
				"operationId": "GetTable",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTable",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITable"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTable",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Table"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tableId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/info": {
			"post": {
				"operationId": "UserInfo",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user": {
			"post": {
				"operationId": "UserSignup",
				"responses": {
					"201": {
						"description": "Create Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAlluser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"201": {
						"description": "Update Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUser"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/login": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IuserLogin"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}